geodns-explorer/anchor/requirements.txt
fastapi
uvicorn[standard]
python-multipart
dnspython>=2.4.0
